enterprise-solutions/rfp-scraper-enterprise/requirements.txt
# Core dependencies
aiohttp==3.8.5
asyncio
beautifulsoup4==4.12.2
cryptography==41.0.4
lxml==4.9.3
prometheus-client==0.17.1
psutil==5.9.5
PyYAML==6.0.1
requests==2.31.0
retrying==1.3.4
backoff==2.2.1
schedule==1.2.0

# Machine Learning dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
joblib>=1.2.0

# Optional dependencies for enhanced features
python-dotenv>=0.19.0
click>=8.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
flake8>=5.0.0